// src/utils/enums/Status.ts
export enum Status {
	ACTIVE = "ACTIVE",
	INACTIVE = "INACTIVE",
	TERMINATED = "TERMINATED",
	TERMINATE = "TERMINATE"
}

export enum PaymentStatusEnum {
	PURCHASE_INITIATED = "PURCHASE_INITIATED",
	PAYMENT_DONE = "PAYMENT_DONE",
	PAYMENT_FAILED = "PAYMENT_FAILED"
}

export enum RequestStatusEnum {
	REJECTED = "REJECTED",
	EXPIRED = "EXPIRED",
	REQUESTED = "REQUESTED",
	APPROVED = "APPROVED",
	VERIFICATION_PENDING = "VERIFICATION_PENDING",
	NEW_PARTICIPANT = "NEW_PARTICIPANT"
}
// ...
export enum InviteeTypeEnum {
	PAYMENT_REQUEST = "PAYMENT_REQUEST",
	PARTICIPANT = "PARTICIPANT"
}

export enum PayeeTypeEnum {
	SELF = "SELF",
	OTHER = "OTHER"
}

export enum ParticipantUserType {
	PRIMARY = "PRIMARY",
	PARTICIPANT = "PARTICIPANT"
}

export enum UserProfileStatusEnum {
	SUBSCRIPTION_PENDING = "SUBSCRIPTION_PENDING",
	REFDOC_PENDING = "REFDOC_PENDING",
	LEASE_DATA_PENDING = "LEASE_DATA_PENDING",
	PLAID_PENDING = "PLAID_PENDING",
	HOME = "HOME",
	INVITED_PLAID_PENDING = "INVITED_PLAID_PENDING",
	INVITED_HOME = "INVITED_HOME",
	INVITED_LEASE_DATA_PENDING = "INVITED_LEASE_DATA_PENDING",
	PAYMENT_DETAIL_PENDING = "PAYMENT_DETAIL_PENDING"
}
// ...
export enum YNStatusEnum {
	YES = "Y",
	NO = "N"
}

export enum DisplayYesNoEnum {
	YES = "Yes",
	NO = "No"
}